% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{biosignMultiDataSet-class}
\alias{biosignMultiDataSet-class}
\title{Class "biosignMultiDataSet"}
\description{
An S4 class to store the the biosign objects generated by the application of the
'biosign' method to a MultiDataSet instance
}
\section{Slots}{

\describe{
\item{\code{biosignLs}}{List: List of instances from the 'biosign' class corresponding to
the models built on each ExpresssionSet}
}}

\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("biosignMultiDataSet", ...)} or by applying the \code{biosign} function to
a MultiDataSet instance
}

\examples{
In progress
}
\seealso{
\code{\link{biosign}}
}
