% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,biosign-method}
\alias{plot,biosign-method}
\alias{plot.biosign}
\title{Plot method for 'biosign' signature objects}
\usage{
\S4method{plot}{biosign}(x, y, tierMaxC = "S", typeC = c("tier",
  "boxplot")[1], file.pdfC = NULL, .sinkC = NULL, fig.pdfC = NA,
  info.txtC = NA, ...)
}
\arguments{
\item{x}{An S4 object of class \code{biosign}, created by the \code{biosign}
function.}

\item{y}{Currently not used.}

\item{tierMaxC}{Character: Maximum level of tiers to display: Either 'S' and
'A', (for boxplot), or also 'B', 'C', 'D', and 'E' (for tiers) by decreasing
number of selections}

\item{typeC}{Character: Plot type; either 'tier' [default] displaying the
comparison of signatures up to the selected 'tierMaxC' or 'boxplot' showing
the individual boxplots of the features selected by all the classifiers}

\item{file.pdfC}{Character: deprecated; use the 'fig.pdfC' argument instead}

\item{.sinkC}{Character: deprecated; use the 'info.txtC' argument instead}

\item{fig.pdfC}{Figure filename (e.g. in case of batch mode) ending with
'.pdf'; default is NA (no saving; displaying instead)}

\item{info.txtC}{Character: Report filename for R output diversion [default =
NA: no diversion]}

\item{...}{Currently not used.}
}
\value{
A plot is created on the current graphics device.
}
\description{
Displays classifier tiers or individual boxplots from selected features
}
\examples{

## loading the diaplasma dataset

data(diaplasma)
attach(diaplasma)

## restricting to a smaller dataset for this example

featureSelVl <- variableMetadata[, "mzmed"] >= 490 & variableMetadata[, "mzmed"] < 500
dataMatrix <- dataMatrix[, featureSelVl]
variableMetadata <- variableMetadata[featureSelVl, ]

## signature selection for all 3 classifiers
## a bootI = 5 number of bootstraps is used for this example
## we recommend to keep the default bootI = 50 value for your analyzes

set.seed(123)
diaSign <- biosign(dataMatrix, sampleMetadata[, "type"], bootI = 5)

## individual boxplot of the selected signatures

plot(diaSign, typeC = "boxplot")

detach(diaplasma)

}
\author{
Philippe Rinaudo and Etienne Thevenot (CEA)
}
